+++ src/components/main/servo.rc
@@ -81,7 +81,6 @@ pub mod layout {
     pub mod context;
     pub mod display_list_builder;
     pub mod float_context;
-    pub mod float;
     pub mod flow;
     pub mod layout_task;
     pub mod inline;
